# Refactor to improve performance w/ weak map, hoisted regex