--- a/lib/core.js
+++ b/lib/core.js
@@ -13,6 +13,15 @@
  * @typedef {CoreOptions & FormatOptions & import('./util/format-smart.js').FormatSmartOptions} CoreWithFormatOptions
  */
 
+const defaultSubsetRegex = /["&'<>`]/g
+const surrogatePairsRegex = /[\uD800-\uDBFF][\uDC00-\uDFFF]/g
+// eslint-disable-next-line no-control-regex, unicorn/no-hex-escape
+const controlCharactersRegex = /[\x01-\t\v\f\x0E-\x1F\x7F\x81\x8D\x8F\x90\x9D\xA0-\uFFFF]/g
+const regexEscapeRegex = /[|\\{}()[\]^$+*?.]/g
+
+/** @type {WeakMap<string[], RegExp>} */
+const subsetToRegexCache = new WeakMap()
+
 /**
  * Encode certain characters in `value`.
  *
@@ -22,7 +31,9 @@
  */
 export function core(value, options) {
   value = value.replace(
-    options.subset ? charactersToExpression(options.subset) : /["&'<>`]/g,
+    options.subset
+      ? charactersToExpression(options.subset)
+      : defaultSubsetRegex,
     basic
   )
 
@@ -33,14 +44,10 @@ export function core(value, options) {
   return (
     value
       // Surrogate pairs.
-      .replace(/[\uD800-\uDBFF][\uDC00-\uDFFF]/g, surrogate)
+      .replace(surrogatePairsRegex, surrogate)
       // BMP control characters (C0 except for LF, CR, SP; DEL; and some more
       // non-ASCII ones).
-      .replace(
-        // eslint-disable-next-line no-control-regex, unicorn/no-hex-escape
-        /[\x01-\t\v\f\x0E-\x1F\x7F\x81\x8D\x8F\x90\x9D\xA0-\uFFFF]/g,
-        basic
-      )
+      .replace(controlCharactersRegex, basic)
   )
 
   /**
@@ -78,13 +85,18 @@ export function core(value, options) {
  * @returns {RegExp}
  */
 function charactersToExpression(subset) {
+  const cached = subsetToRegexCache.get(subset)
+  if (cached) return cached
+
   /** @type {Array<string>} */
   const groups = []
   let index = -1
 
   while (++index < subset.length) {
-    groups.push(subset[index].replace(/[|\\{}()[\]^$+*?.]/g, '\\$&'))
+    groups.push(subset[index].replace(regexEscapeRegex, '\\$&'))
   }
 
-  return new RegExp('(?:' + groups.join('|') + ')', 'g')
+  const regex = new RegExp('(?:' + groups.join('|') + ')', 'g')
+  subsetToRegexCache.set(subset, regex)
+  return regex
 }