--- a/lib/core.js
+++ b/lib/core.js
@@ -1,6 +1,6 @@
 /**
  * @typedef CoreOptions
- * @property {Array<string>} [subset=[]]
+ * @property {ReadonlyArray<string>} [subset=[]]
  *   Whether to only escape the given subset of characters.
  * @property {boolean} [escapeOnly=false]
  *   Whether to only escape possibly dangerous characters.
@@ -15,11 +15,12 @@
 
 const defaultSubsetRegex = /["&'<>`]/g
 const surrogatePairsRegex = /[\uD800-\uDBFF][\uDC00-\uDFFF]/g
-// eslint-disable-next-line no-control-regex, unicorn/no-hex-escape
-const controlCharactersRegex = /[\x01-\t\v\f\x0E-\x1F\x7F\x81\x8D\x8F\x90\x9D\xA0-\uFFFF]/g
+const controlCharactersRegex =
+  // eslint-disable-next-line no-control-regex, unicorn/no-hex-escape
+  /[\x01-\t\v\f\x0E-\x1F\x7F\x81\x8D\x8F\x90\x9D\xA0-\uFFFF]/g
 const regexEscapeRegex = /[|\\{}()[\]^$+*?.]/g
 
-/** @type {WeakMap<string[], RegExp>} */
+/** @type {WeakMap<ReadonlyArray<string>, RegExp>} */
 const subsetToRegexCache = new WeakMap()
 
 /**
@@ -32,7 +33,7 @@ const subsetToRegexCache = new WeakMap()
 export function core(value, options) {
   value = value.replace(
     options.subset
-      ? charactersToExpression(options.subset)
+      ? charactersToExpressionCached(options.subset)
       : defaultSubsetRegex,
     basic
   )
@@ -81,13 +82,29 @@ export function core(value, options) {
 }
 
 /**
- * @param {Array<string>} subset
+ * A wrapper function that caches the result of `charactersToExpression` with a WeakMap.
+ * This can improve performance when tooling calls `charactersToExpression` repeatedly
+ * with the same subset.
+ *
+ * @param {ReadonlyArray<string>} subset
+ * @returns {RegExp}
+ */
+function charactersToExpressionCached(subset) {
+  let cached = subsetToRegexCache.get(subset)
+
+  if (!cached) {
+    cached = charactersToExpression(subset)
+    subsetToRegexCache.set(subset, cached)
+  }
+
+  return cached
+}
+
+/**
+ * @param {ReadonlyArray<string>} subset
  * @returns {RegExp}
  */
 function charactersToExpression(subset) {
-  const cached = subsetToRegexCache.get(subset)
-  if (cached) return cached
-
   /** @type {Array<string>} */
   const groups = []
   let index = -1
@@ -96,7 +113,5 @@ function charactersToExpression(subset) {
     groups.push(subset[index].replace(regexEscapeRegex, '\\$&'))
   }
 
-  const regex = new RegExp('(?:' + groups.join('|') + ')', 'g')
-  subsetToRegexCache.set(subset, regex)
-  return regex
+  return new RegExp('(?:' + groups.join('|') + ')', 'g')
 }