- This PR hoists regexes as top-level variables to prevent constantly creating them.
- It also implements caching for `charactersToExpression` to prevent repeatedly created regexes dynamically.

### Benchmark

During my benchmarking of a project using `toHtml` heavily from `hast-util-to-html`, `toHtml` took about 6.06s in execution time. This PR drops it to 2.38s. `toHtml` uses this package which is hot code during its `serializeAttribute` call. Examples: https://github.com/search?q=repo%3Asyntax-tree%2Fhast-util-to-html%20subset%3A&type=code

### Notes on caching

The WeakMap caching for `charactersToExpression` means that if the array is mutated, the regex won't be re-built but I think this is a fine caveat that shouldn't be relied upon. If I remove the WeakMap caching, the benchmark only drops to 5.52s.

In `hast-util-to-html`, there are two places (`lib/handle/text.js` and `lib/handle/comment.js`) that can yet to take advantage of the caching. If this PR is accepted, I can update `hast-util-to-html` to do so, and I expect another 1s savings, bringing down to around 1.4s for `toHtml`.