@wooorm 2024-04-03
> array is mutated

I’m :+1: to using `ReadonlyArray` in the types. I don’t consider it a breaking change either.