@bluwy 2024-04-03
Updated! 

> Also, maybe good for reference, as this is 2 things in 1: do you have the separate numbers for a) weak map, b) hoisted regexes?

For weak map only, I measured `toHtml` and it dropped to 2.69s. For hoisted regexes only, it dropped to 5.52s. If both are applied, it dropped to 2.38s. (NOTE: Initial timing is 6.06s)

So it seems like hoisting the regex has only a small gain, but I think it's because `toHtml` always passes a `subset`, so it hits  this part only: https://github.com/wooorm/stringify-entities/blob/632a2088c833b42453d19dfbf49d1d850b7f8e80/lib/core.js#L24-L31

Skipping most of the regex work after it. Other usages for `stringify-entities` should have a bigger impact.