@wooorm 2024-04-03
Thanks! Released in [`4.0.4`](https://github.com/wooorm/stringify-entities/releases/tag/4.0.4)!